refactor location of mysqlcheck
